requirements.txt
selenium
requests
beautifulsoup4
googletrans==4.0.0rc1
python-dotenv
pytest
browserstack-local 